Allow caller to specify pegdown extensions used
Currently the generator hard-codes all pegdown extensions to on. This is
annoying as e.g. hard line breaks are a bad fit for documentation.

This changes that to a default, and allows a caller to pass through a pegdown
extension bit-mask integer to customise the extensions to their own taste.